+++ package.json
@@ -3,10 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "prop-types": "^15.8.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
-    "react-scripts": "5.0.1",
     "styled-components": "^5.3.5"
   },
   "scripts": {
@@ -64,6 +62,9 @@
     "eslint-plugin-react-hooks": "^4.4.0",
     "eslint-plugin-storybook": "^0.5.11",
     "prettier": "^2.6.2",
+    "prop-types": "^15.8.1",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
     "webpack": "^5.72.0"
   },
   "homepage": "https://al-bur.github.io/react-payments/"
