--- src/index.css
@@ -1,10 +1,0 @@
-body {
-  background: #e5e5e5;
-  display: flex;
-  font-family: 'Roboto', 'Malgun Gothic';
-  justify-content: center;
-}
-
-::placeholder {
-  font-family: 'Roboto', 'Malgun Gothic' !important;
-}
