+++ src/pages/AddCardComplete.jsx
@@ -1,0 +1,5 @@
+function AddCardComplete() {
+  return <div>완료</div>;
+}
+
+export default AddCardComplete;
